+++ src/main/java/de/rub/nds/scanner/core/config/ExecutorConfig.java
@@ -9,8 +9,8 @@
 package de.rub.nds.scanner.core.config;
 
 import com.beust.jcommander.Parameter;
-import de.rub.nds.scanner.core.constants.ProbeType;
-import de.rub.nds.scanner.core.constants.ScannerDetail;
+import de.rub.nds.scanner.core.probe.ProbeType;
+
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
