+++ src/main/java/de/rub/nds/scanner/core/config/ScannerDetail.java
@@ -6,7 +6,7 @@
  * Licensed under Apache License, Version 2.0
  * http://www.apache.org/licenses/LICENSE-2.0.txt
  */
-package de.rub.nds.scanner.core.constants;
+package de.rub.nds.scanner.core.config;
 
 public enum ScannerDetail implements Comparable<ScannerDetail> {
     ALL(100),
