+++ src/main/java/de/rub/nds/scanner/core/probe/AnalyzedProperty.java
@@ -6,10 +6,17 @@
  * Licensed under Apache License, Version 2.0
  * http://www.apache.org/licenses/LICENSE-2.0.txt
  */
-package de.rub.nds.scanner.core.constants;
+package de.rub.nds.scanner.core.probe;
 
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
+@XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
-public interface AnalyzedPropertyCategory {}
+public interface AnalyzedProperty {
+
+    AnalyzedPropertyCategory getCategory();
+
+    String getName();
+}
