+++ src/main/java/de/rub/nds/scanner/core/probe/ScannerProbe.java
@@ -8,7 +8,6 @@
  */
 package de.rub.nds.scanner.core.probe;
 
-import de.rub.nds.scanner.core.constants.*;
 import de.rub.nds.scanner.core.passive.StatsWriter;
 import de.rub.nds.scanner.core.probe.requirements.Requirement;
 import de.rub.nds.scanner.core.probe.result.*;
