+++ src/main/java/de/rub/nds/scanner/core/report/AnalyzedPropertyTextEncoder.java
@@ -8,7 +8,7 @@
  */
 package de.rub.nds.scanner.core.report;
 
-import de.rub.nds.scanner.core.constants.AnalyzedProperty;
+import de.rub.nds.scanner.core.probe.AnalyzedProperty;
 import java.util.HashMap;
 
 public class AnalyzedPropertyTextEncoder extends Encoder<AnalyzedProperty> {
