+++ src/main/java/de/rub/nds/scanner/core/report/ReportPrinter.java
@@ -8,8 +8,8 @@
  */
 package de.rub.nds.scanner.core.report;
 
-import de.rub.nds.scanner.core.constants.AnalyzedProperty;
-import de.rub.nds.scanner.core.constants.ScannerDetail;
+import de.rub.nds.scanner.core.probe.AnalyzedProperty;
+import de.rub.nds.scanner.core.config.ScannerDetail;
 
 public abstract class ReportPrinter<ReportT extends ScanReport> {
