+++ src/main/java/de/rub/nds/scanner/core/report/rating/PropertyResultRatingInfluencer.java
@@ -8,7 +8,7 @@
  */
 package de.rub.nds.scanner.core.report.rating;
 
-import de.rub.nds.scanner.core.constants.AnalyzedProperty;
+import de.rub.nds.scanner.core.probe.AnalyzedProperty;
 import de.rub.nds.scanner.core.probe.result.TestResult;
 import de.rub.nds.scanner.core.probe.result.TestResults;
 import jakarta.xml.bind.annotation.XmlAccessType;
