+++ src/main/java/de/rub/nds/scanner/core/report/rating/RatingInfluencers.java
@@ -8,7 +8,7 @@
  */
 package de.rub.nds.scanner.core.report.rating;
 
-import de.rub.nds.scanner.core.constants.AnalyzedProperty;
+import de.rub.nds.scanner.core.probe.AnalyzedProperty;
 import de.rub.nds.scanner.core.probe.result.TestResult;
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
