+++ src/main/java/de/rub/nds/scanner/core/report/rating/ScoreReport.java
@@ -8,7 +8,7 @@
  */
 package de.rub.nds.scanner.core.report.rating;
 
-import de.rub.nds.scanner.core.constants.AnalyzedProperty;
+import de.rub.nds.scanner.core.probe.AnalyzedProperty;
 import java.util.LinkedHashMap;
 
 public class ScoreReport {
