+++ src/test/java/de/rub/nds/scanner/core/probe/requirements/ProbeRequirementTest.java
@@ -13,7 +13,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import de.rub.nds.scanner.core.TestReport;
-import de.rub.nds.scanner.core.constants.ProbeType;
+import de.rub.nds.scanner.core.probe.ProbeType;
 import org.junit.jupiter.api.Test;
 
 public class ProbeRequirementTest {
